Delete legacy toksearch_d3d.fdp subpackage; rename module to fdp.py
Drops cli.py, environment.py, skills.py (now owned by the fdp package).
Drops test_fdp_query.py and test_fdp_environment.py (tests moved to
the fdp repo). Renames the transition module from _fdp_new.py to its
final home at toksearch_d3d.fdp; updates imports and entry-point path
to match.
