--- toksearch_d3d/fdp/__init__.py
@@ -1,3 +1,0 @@
-from .environment import setup_environment
-
-__all__ = ["setup_environment"]
