--- toksearch_d3d/fdp/cli.py
@@ -1,388 +1,0 @@
-from pathlib import Path
-import argparse
-import shlex
-import shutil
-import sys
-import os
-import subprocess
-
-from XRootD import client
-from XRootD.client.flags import DirListFlags, StatInfoFlags
-
-from .skills import _parse_skill_md
-from .environment import (
-    DEFAULT_CONFIG,
-    FDP_ROOT,
-    ORIGIN_SERVER,
-    setup_environment,
-)
-
-
-class FdpFileSystem:
-    def __init__(self, server: str):
-        self.server = server
-        self.xrd_fs = client.FileSystem(server)
-
-    def ls(self, path: str | Path, dirs_only: bool = False) -> list[Path]:
-        _, listings = self.xrd_fs.dirlist(str(path), DirListFlags.STAT)
-
-        if not listings:
-            return []
-
-        if dirs_only:
-            listings = [l for l in listings if l.statinfo.flags & StatInfoFlags.IS_DIR]
-
-        return [Path(l.name) for l in listings]
-
-
-##############################################################################
-#
-# SKILLS BACKENDS
-#
-##############################################################################
-
-
-class ClaudeBackend:
-    """Claude Code — ~/.claude/skills/<name>/SKILL.md"""
-
-    name = "claude"
-
-    @property
-    def dest_root(self):
-        return Path.home() / ".claude" / "skills"
-
-    def is_detected(self):
-        return (Path.home() / ".claude").exists()
-
-    def install_skill(self, skill_dir: Path, force: bool) -> str:
-        dest = self.dest_root / skill_dir.name
-        if dest.exists() and not force:
-            return "skipped"
-        if dest.exists():
-            shutil.rmtree(dest)
-        shutil.copytree(skill_dir, dest)
-        return "installed"
-
-    def is_skill_installed(self, skill_name: str) -> bool:
-        return (self.dest_root / skill_name).exists()
-
-
-class CursorBackend:
-    """Cursor IDE — ~/.cursor/rules/fdp-<name>.mdc"""
-
-    name = "cursor"
-
-    @property
-    def dest_root(self):
-        return Path.home() / ".cursor" / "rules"
-
-    def is_detected(self):
-        return (Path.home() / ".cursor").exists()
-
-    def install_skill(self, skill_dir: Path, force: bool) -> str:
-        skill_md = skill_dir / "SKILL.md"
-        if not skill_md.exists():
-            return "skipped"
-        fm, body = _parse_skill_md(skill_md)
-        description = fm.get("description", skill_dir.name)
-        dest_file = self.dest_root / f"fdp-{skill_dir.name}.mdc"
-        if dest_file.exists() and not force:
-            return "skipped"
-        self.dest_root.mkdir(parents=True, exist_ok=True)
-        content = f"---\ndescription: {description}\nglobs: \nalwaysApply: false\n---\n\n{body}"
-        dest_file.write_text(content)
-        return "installed"
-
-    def is_skill_installed(self, skill_name: str) -> bool:
-        return (self.dest_root / f"fdp-{skill_name}.mdc").exists()
-
-
-class CodexBackend:
-    """OpenAI Codex CLI — ~/.codex/instructions.md (section-per-skill)"""
-
-    name = "codex"
-
-    @property
-    def dest_root(self):
-        return Path.home() / ".codex"
-
-    @property
-    def _instructions_file(self):
-        return self.dest_root / "instructions.md"
-
-    def is_detected(self):
-        return (Path.home() / ".codex").exists()
-
-    def _markers(self, skill_name):
-        return f"<!-- fdp-skill:{skill_name} -->", f"<!-- /fdp-skill:{skill_name} -->"
-
-    def is_skill_installed(self, skill_name: str) -> bool:
-        if not self._instructions_file.exists():
-            return False
-        start, _ = self._markers(skill_name)
-        return start in self._instructions_file.read_text()
-
-    def install_skill(self, skill_dir: Path, force: bool) -> str:
-        import re
-        skill_md = skill_dir / "SKILL.md"
-        if not skill_md.exists():
-            return "skipped"
-        _, body = _parse_skill_md(skill_md)
-        start, end = self._markers(skill_dir.name)
-        section = f"{start}\n{body.rstrip()}\n{end}"
-        current = self._instructions_file.read_text() if self._instructions_file.exists() else ""
-        if start in current:
-            if not force:
-                return "skipped"
-            pattern = re.escape(start) + r".*?" + re.escape(end)
-            current = re.sub(pattern, section, current, flags=re.DOTALL)
-        else:
-            current = current.rstrip("\n") + ("\n\n" if current else "") + section + "\n"
-        self._instructions_file.write_text(current)
-        return "installed"
-
-
-BACKENDS = {
-    "claude": ClaudeBackend(),
-    "cursor": CursorBackend(),
-    "codex":  CodexBackend(),
-}
-
-
-##############################################################################
-#
-# CLI SUB COMMMANDS
-#
-##############################################################################
-
-
-def do_skills(args):
-    import toksearch
-    import toksearch_d3d
-
-    skill_sources = [
-        Path(toksearch.__file__).parent / "skills",
-        Path(toksearch_d3d.__file__).parent / "skills",
-    ]
-    skill_dirs = [
-        d
-        for source in skill_sources if source.exists()
-        for d in sorted(source.iterdir()) if d.is_dir()
-    ]
-
-    backend_arg = getattr(args, "backend", "claude")
-    if backend_arg == "all":
-        backends = [b for b in BACKENDS.values() if b.is_detected()]
-        if not backends:
-            print("No supported coding assistant tool detected.")
-            return
-    elif backend_arg in BACKENDS:
-        backends = [BACKENDS[backend_arg]]
-    else:
-        print(f"Unknown backend '{backend_arg}'. Choose from: {', '.join(BACKENDS)}, all")
-        sys.exit(1)
-
-    if args.skills_command == "list":
-        for backend in backends:
-            print(f"[{backend.name}]")
-            for d in skill_dirs:
-                status = "installed" if backend.is_skill_installed(d.name) else "not installed"
-                print(f"  {d.name}  [{status}]")
-        return
-
-    force = getattr(args, "force", False)
-    for backend in backends:
-        print(f"\n[{backend.name}] Installing to {backend.dest_root}")
-        backend.dest_root.mkdir(parents=True, exist_ok=True)
-        installed = skipped = 0
-        for skill_dir in skill_dirs:
-            result = backend.install_skill(skill_dir, force)
-            if result == "installed":
-                print(f"  install  {skill_dir.name}")
-                installed += 1
-            else:
-                print(f"  skip     {skill_dir.name}  (use --force to overwrite)")
-                skipped += 1
-        print(f"  {installed} installed, {skipped} skipped")
-
-
-def do_run(args):
-    passthrough_args = args.command_args
-
-    if args.debug:
-        command_str = " ".join(passthrough_args)
-        print(f"Running command: {command_str}")
-        print("With environment vars:")
-        for k, v in os.environ.items():
-            print(f"{k}: {v}")
-
-    comm = passthrough_args
-    result = subprocess.run(comm, env=os.environ)
-    sys.exit(result.returncode)
-
-
-def do_env(args):
-    for key, value in DEFAULT_CONFIG.items():
-        print(f"export {key}={shlex.quote(value)}")
-    bearer_token = os.environ.get("BEARER_TOKEN", "")
-    if bearer_token:
-        print(f"export BEARER_TOKEN={shlex.quote(bearer_token)}")
-
-
-def do_ls(args):
-    fs = FdpFileSystem(ORIGIN_SERVER)
-
-    if args.path == "/":
-        listing = [Path(FDP_ROOT).name]
-    else:
-        listing = fs.ls(args.path)
-
-    if listing:
-        for l in listing:
-            print(l)
-    else:
-        print("No such file or directory")
-        sys.exit(1)
-
-
-def _build_llm_cmd(subcommand: str, passthrough_args: list[str]) -> list[str]:
-    """Construct argv for the toksearch.llm CLI delegate."""
-    cmd = [sys.executable, "-m", "toksearch.llm.cli", subcommand]
-    # Default backend: amsc, preserves existing `fdp query` behavior so
-    # users with ~/amsc_api_key see no change.  The user can override with
-    # --backend on the fdp side.
-    if "--backend" not in passthrough_args:
-        cmd.extend(["--backend", "amsc"])
-    cmd.extend(passthrough_args)
-    return cmd
-
-
-def do_query(args):
-    # Re-exec into python -m toksearch.llm.cli query.  The exec replaces
-    # this Python process so libfdpio/XRootD env vars set by
-    # setup_environment() are honored at the new process's C-library load
-    # time (subprocess vs re-exec is equivalent here for env block; re-exec
-    # is simpler and avoids subprocess wait/return-code plumbing).
-    passthrough = [args.query]
-    if args.backend:
-        passthrough.extend(["--backend", args.backend])
-    if args.model:
-        passthrough.extend(["--model", args.model])
-    if args.max_iterations is not None:
-        passthrough.extend(["-n", str(args.max_iterations)])
-    cmd = _build_llm_cmd("query", passthrough)
-    os.execvpe(cmd[0], cmd, os.environ)
-
-
-def do_chat(args):
-    passthrough = []
-    if args.backend:
-        passthrough.extend(["--backend", args.backend])
-    if args.model:
-        passthrough.extend(["--model", args.model])
-    if args.max_iterations is not None:
-        passthrough.extend(["-n", str(args.max_iterations)])
-    cmd = _build_llm_cmd("chat", passthrough)
-    os.execvpe(cmd[0], cmd, os.environ)
-
-
-##############################################################################
-#
-# MAIN
-#
-##############################################################################
-def main():
-    parser = argparse.ArgumentParser(
-        description="CLI interface for the Fusion Data Platform"
-    )
-    # Define any wrapper-specific options here if needed
-    parser.add_argument("--bearer-token", "-t", type=str, default="")
-    parser.add_argument("--debug", action="store_true", help="Print debug info")
-
-    subparsers = parser.add_subparsers(dest="command")
-
-    run_parser = subparsers.add_parser(
-        "run", help="Run a command with access to FDP data"
-    )
-    run_parser.add_argument(
-        "command_args",
-        nargs=argparse.REMAINDER,
-        help="Command and args to pass through",
-    )
-    run_parser.set_defaults(func=do_run)
-
-    env_parser = subparsers.add_parser(
-        "env", help="Print environment variables for shell eval"
-    )
-    env_parser.set_defaults(func=do_env)
-
-    ls_parser = subparsers.add_parser("ls", help="List files on the FDP")
-    ls_parser.add_argument(
-        "--dirs-only", "-d", action="store_true", help="Only show subdirectories"
-    )
-    ls_parser.add_argument(
-        "path", type=str, help="The path whose contents will be listed"
-    )
-    ls_parser.set_defaults(func=do_ls)
-
-    skills_parser = subparsers.add_parser("skills", help="Manage AI coding assistant skills for FDP")
-    skills_sub = skills_parser.add_subparsers(dest="skills_command")
-
-    list_p = skills_sub.add_parser("list", help="List FDP skills and their installation status")
-    list_p.add_argument(
-        "--backend", default="claude",
-        help="Target tool: claude, cursor, codex, or all (default: claude)"
-    )
-
-    install_p = skills_sub.add_parser("install", help="Install FDP skills for a coding assistant")
-    install_p.add_argument(
-        "--backend", default="claude",
-        help="Target tool: claude, cursor, codex, or all (default: claude)"
-    )
-    install_p.add_argument(
-        "--force", "-f", action="store_true", help="Overwrite already-installed skills"
-    )
-
-    skills_parser.set_defaults(func=do_skills)
-
-    def _add_llm_args(parser):
-        parser.add_argument(
-            "--backend", default=None,
-            help="Backend / preset name (default: amsc). Options: amsc, "
-                 "anthropic, openai, claude-max, or any user preset.")
-        parser.add_argument(
-            "--model", default=None,
-            help="Override the preset's default model.")
-        parser.add_argument(
-            "-n", "--max-iterations", type=int, default=None,
-            help="Cap on tool-call rounds per turn.")
-
-    query_parser = subparsers.add_parser(
-        "query",
-        help="Run a one-shot natural-language query against TokSearch",
-    )
-    query_parser.add_argument(
-        "query",
-        type=str,
-        help="Natural-language query (quote it on the shell)",
-    )
-    _add_llm_args(query_parser)
-    query_parser.set_defaults(func=do_query)
-
-    chat_parser = subparsers.add_parser(
-        "chat",
-        help="Interactive conversational query against TokSearch",
-    )
-    _add_llm_args(chat_parser)
-    chat_parser.set_defaults(func=do_chat)
-
-    args = parser.parse_args()
-
-    setup_environment(bearer_token=args.bearer_token or None)
-
-    # Now run it
-    args.func(args)
-
-
-if __name__ == "__main__":
-    main()
